Moved pagination and algolia search hooks out of businesses folder
Figured by this stage they had gotten big enough to deserve their own
folder. Also pagination can be used by the allies page as well as that
is all the same stuff.
